// src/models/films/films.ts
export class FilmsModel {
    name: string;
    note: string;
	title: string;
	episode_id: number;
	opening_crawl: string;
	director: string;
	release_date: any;
	created: any;
	edited: any;
	url: string;
	characters: Array <string> = [];
	planets: Array <string> = [];
	starships: Array <string> = [];
	vehicles: Array <string> = [];
	species: Array <string> = [];


}